Add `post_glob' parameter for `register()'
Sometimes, you want to be able to perform a glob but exclude some of the
matched files. The `bsd_glob()' function from File::Glob doesn't allow
for that.

So now, you can supply a coderef that does all that excluding for you.
Predefined functions are `remove_tilde' for removing files named "*~"
and `remove_hashes' for removing files whose basename matches "#*#".

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>