First set of changes for unifying dewi into one tool
Here's the thing: I should have never depended on ‘make’ in the first
place. I'm basically only using it to call the dewi script and because
the "make deploy" looks pretty. But there's no feature in make that the
basic dewi features actually need. And then you just end up with a bunch
of downsides, like:

  - Superfluous dependency, that a target system has to satisfy

  - Proper portable make code is actually not that easy, and I always
    get issues on OpenBSD systems for example.

  - Using make required a Makefile in each dewified sub-directories
    to exist. Which means that you pollute the sub-directory's
    version control history with changes to that file each time dewi
    updates it.

  - If someone wants to use make as part of their deployment system,
    they can use it on top of a make-less dewi-system at will.

  - Needlessly complicates the whole system.

You can think of more. It was just wrong to use ‘make’.

And then there is the case of keeping the code for dewi in the dot-dewi
sub-directory. The idea was to ensure, that the deployment process
doesn't ever break due to changes in dewi that might be in updated
system version. That's still a moot point however, since nobody keeps
someone from installing a fixed version of dewi into ‘~/bin’ and never
ever change it.

So the plan is to unify ALL of dewi into the CLI script. That'll largely
simplify the whole system and thin out the dot-dewi sub-directory.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>