Port options and user settings to Dewi::Storage::Options
Now user settings and options are cleanly separated into global and
per-package scope, meaning that the user is able to set values in
.dewi/Dewifile to be default values, without breaking if called from
within a package subdirectory or if called from the dewi root directory
with one or more packages to deploy as arguments.

This still lacks documentation and read/write access to default values
from within packages (to be used with caution of course). But this is a
major step towards version 0.3.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>